Move all the island stuff to `mod island`.
